# perf(install): keep parsed npm package info in memory cache